Update NavBar for clerk and final design
This commit updates the navbar to match the figma final design and
integrates clerk: IF the user is signed out, there is a sign in button
at the top of the page redirecting to the sign-in page. Otherwise, it
displays the userprofile which on click will redirect to the account
page. This uses clerk's built in UserButton which pre-populates an
account page.